chore(charset issue): Resolution for issue 6102
Resolution for the issue https://github.com/psf/requests/issues/6102

Now, If a username and password are passed already encoded, they will not be affected by being expected as a string. In this case, the function will encode the string-formatted attributes into bytes using the latin-1 charset by default due to convention.